binutils-config: fix native link removal by cross-compilers, bug #724454
commit 36eba05752348258a7 ("binutils-config: add build-time
and runtime switches to disable native symlinks").

Added symlink deletion for USE_NATIVE_LINKS=no mode.
It had a bug: it unconditionally deleted native symlinks
if binutils-config is called on a cross-binutils installation.

I missed the case because I assumed that tool names
in binutils' private directory match gcc' private directory.

The change is not to perform symlink cleanup for unprefixed
binaries for cross-binutils.

Reported-by: Ionen Wolkens
Bug: https://bugs.gentoo.org/724454
Signed-off-by: Sergei Trofimovich <slyfox@gentoo.org>